Added custom ListView and row item for MainFragment

+++ app/src/main/java/com/hagarsoft/weatherapp/data/Coord.java
@@ -1,6 +1,6 @@
 package com.hagarsoft.weatherapp.data;
 
 public class Coord {
-    public float lat;
-    public float lon;
+    public double lat;
+    public double lon;
 }
